requirements.txt
openai
sounddevice
pynput
pyaudio
pydub
langchain
python-dotenv
